Abstract scraping logic; rewrite Playlist to use scraping
While the API was faster and safer for YoutubePlaylist, it also
wasn't able to load private playlists without oauth, which I didn't
want to have to introduce. Switching to scraping simplifies the use
of this library as well, since you don't need to generate an API
key to access playlists.