Enable dumpio if debug() is enabled; disable navigation timeout
Having dumpio is useful when debugging issues like "the whole browser
crashes when getting to the 2FA step." It seems like having a Security
Key on your Google Account triggers some sort of "(os/kern) failure" and
removing the key temporarily causes it to work just fine. This happens
even when controlling Chrome instead of Chromium, for some reason.